Adjusted lightning run parameters
Removed requirement to use the 'duck' key and instead made it so that the story engine stall for the lightning run happens just like in the film by manipulating the alarm clock and changing the restart to automatic. Also corrected issue that was allowing random engine stalls to affect the DeLorean just prior to the lightning run, which could negatively affect player interaction.