Fixes formatting on the XML files, and fixes a horrendous hack.
Seriously, I have never seen a worst hack in my life.
Basically, it added a bunch of stock game wheels to the carcols. I've fixed this by changing the load order of the BTTF DLC from 51, to 85.

It also changes the modShopLabel of a few mods. It's far from perfect, but it helps people who are trying to use the mod in FiveM. And it costs us absolutely nothing.
Note: It doesn't work in LSC. The game has a character limitation there (or something similar? I don't actually know). But it works fine in Rampage Trainer. Probably works in Menyoo and Simple Trainer too.

I also removed suspension mods, as they were set to do nothing before, so they were useless.