Update Plutonium Gauge
Updated handling of Plutonium Chamber gauge to better match its operation in the films.

BTTF2/3: Time Machines with Mr Fusion will no longer trigger the gauge to move at all. Empty light will continue to work as normal.

BTTF1: For Time Machines with the nuclear reactor the gauge will read full scale when the chamber is full, will slowly move down over the course of several minutes if the TFC is on, and when it hits 0 the EMPTY alarm will go off regardless of whether the TFC is on or off. For clarity the 2D EMPTY indicator will appear instantly in 3rd person view on Time Machines with a nuclear reactor.

Also moved the 'all' door option to the top instead of the bottom of the DeLorean Doors Menu.