Fix high alt script crash
If a user attempts to spawn a time machine non-cinematically above an altitude of 2699 meters, it will crash our script. The game has a limit of 2470 meters above which a vehicle's engine will not operate. So we will enforce a hard limit of 2500 meters, and let gravity do the rest.