Lightning Run & Mod update
Since Thunder in LightningRun.cs can now never be null, we need to check against a different item that can either be null or not null during the lightning run experiment time in order for the setup code to work correctly, so now we check against the wire connecting the streetlamps. Also updated the BTTF1 DeLorean to default to lowered rear suspension as this is how it appears in the first film.